Refactor config code in RusticaAgent
