+++ examples/rustica_local_file.toml
@@ -1,37 +1,37 @@
 # The certificate presented to connecting clients
 server_cert = '''
 -----BEGIN CERTIFICATE-----
-MIIBqzCCAVCgAwIBAgIJAIDetXKVHvT9MAoGCCqGSM49BAMCMBsxGTAXBgNVBAMM
-EEVudGVycHJpc2VSb290Q0EwHhcNMjEwOTA5MTg0ODA1WhcNMjMxMjEzMTg0ODA1
+MIIBqjCCAVCgAwIBAgIJAOI2FtcQeixVMAoGCCqGSM49BAMCMBsxGTAXBgNVBAMM
+EEVudGVycHJpc2VSb290Q0EwHhcNMjIwMTIwMDQwMjA2WhcNMjQwNDI0MDQwMjA2
 WjAxMRAwDgYDVQQDDAdydXN0aWNhMRAwDgYDVQQKDAdSdXN0aWNhMQswCQYDVQQG
-EwJDQTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABA4FXg6srQVPFUNRt8O2xJU/
-lOr8sFeTfMfiXk7TS+rdEI0gk7dxtQsrNSwh+3e5UjjJEQkKbkdkhZHuHUvhS1Wj
+EwJDQTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABAINhoFW/5twPqAHLxjFjmns
+lE1jJMJQXmijymZTJxR0DsNZlwvUgNH+WYQFfq4IVMwypVHgyTYJO+lAAPEeyPOj
 ZzBlMDUGA1UdIwQuMCyhH6QdMBsxGTAXBgNVBAMMEEVudGVycHJpc2VSb290Q0GC
-CQDdGlaBb9C6TDAJBgNVHRMEAjAAMAsGA1UdDwQEAwIE8DAUBgNVHREEDTALggls
-b2NhbGhvc3QwCgYIKoZIzj0EAwIDSQAwRgIhAP0YNQmo8G/LlW6lozn+JmcfT9z9
-cNI4mA6YUI2cWtOPAiEArV3Cnypaq8JBP/GCelXiojGEZHgWLaF4ZsCy+ssQjrs=
+CQCRg096sVtP0zAJBgNVHRMEAjAAMAsGA1UdDwQEAwIE8DAUBgNVHREEDTALggls
+b2NhbGhvc3QwCgYIKoZIzj0EAwIDSAAwRQIhAMfjW/PMrA9/cCg6O835sr22ZrNk
+k/lFOODLqAJPbh3+AiAzeCUyrmxT5VTf6uyFoNT8zMoWSi79rudcdgl+32RqMg==
 -----END CERTIFICATE-----
 '''
 
 # The key for the certificate presented to clients
 server_key = '''
 -----BEGIN PRIVATE KEY-----
-MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgv4HTe9TP3vBptisr
-NEO26q5/71dv21JSpSFpwVGtQ3WhRANCAAQOBV4OrK0FTxVDUbfDtsSVP5Tq/LBX
-k3zH4l5O00vq3RCNIJO3cbULKzUsIft3uVI4yREJCm5HZIWR7h1L4UtV
+MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgkTd0C69xWFX9PmVf
+BeD0ySfG+O0e7p7SXR9xo/enbvahRANCAAQCDYaBVv+bcD6gBy8YxY5p7JRNYyTC
+UF5oo8pmUycUdA7DWZcL1IDR/lmEBX6uCFTMMqVR4Mk2CTvpQADxHsjz
 -----END PRIVATE KEY-----
 '''
 
 # The CA certificate clients must have their identities signed by
 client_ca_cert = '''
 -----BEGIN CERTIFICATE-----
-MIIBMjCB2AIJAPWVCGGAtuF3MAoGCCqGSM49BAMCMCExHzAdBgNVBAMMFkVudGVy
-cHJpc2VDbGllbnRSb290Q0EwHhcNMjEwOTA5MTg0ODA1WhcNMzEwOTA3MTg0ODA1
+MIIBMTCB2AIJAL3yJUJ7ShOJMAoGCCqGSM49BAMCMCExHzAdBgNVBAMMFkVudGVy
+cHJpc2VDbGllbnRSb290Q0EwHhcNMjIwMTIwMDQwMjA2WhcNMzIwMTE4MDQwMjA2
 WjAhMR8wHQYDVQQDDBZFbnRlcnByaXNlQ2xpZW50Um9vdENBMFkwEwYHKoZIzj0C
-AQYIKoZIzj0DAQcDQgAEmGjwIaBON7DIrm/tSTTp96WXiovAWoUYwtziWQmqntMr
-GVjQMx5DyGre2cCbMDz2AyJrEDX1ReX+mFODaP+IdTAKBggqhkjOPQQDAgNJADBG
-AiEApHx6eH8yM+PrH87C/KahHCgcInFyLHvO0vVd8+sKUtICIQD8w6LSUcigT0C6
-woKz1ehFp3cKUtt2UEySnvbL/WxDSQ==
+AQYIKoZIzj0DAQcDQgAEmZbAXHUQEXKB/NmHCG0AcjQA0IsBph+jmcFbw9Na58Cv
+PNZtmm8jQ3Q9R8e3faG6gZuXNe60q/Ea6a6jR5UryDAKBggqhkjOPQQDAgNIADBF
+AiA7H67T6QVZq1pBs6MU6+f/4mxYbVkPi/aCqMthRin7qQIhAP3cYrcO8QibwSjx
+QCYQlNrEWT9OVP/akN0OyFDQIYIU
 -----END CERTIFICATE-----
 '''
